streaming/requirements.txt
openai>=1.50.0
google-genai>=0.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
temporalio>=1.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
